requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
pillow
numpy